+++ Samples/macOS-Swift/macOS-Swift/ViewController.swift
@@ -28,6 +28,20 @@ class ViewController: NSViewController {
         // otherwise nil
         print("\(String(describing: eventId))")
     }
+    
+    @IBAction func captureUserFeedback(_ sender: Any) {
+        let error = NSError(domain: "UserFeedbackErrorDomain", code: 0, userInfo: [NSLocalizedDescriptionKey: "This never happens."])
+
+        let eventId = SentrySDK.capture(error: error) { scope in
+            scope.setLevel(.fatal)
+        }
+        
+        let userFeedback = UserFeedack(eventId: eventId)
+        userFeedback.comments = "It broke on macOS-Swift. I don't know why, but this happens."
+        userFeedback.email = "john@me.com"
+        userFeedback.name = "John Me"
+        SentrySDK.capture(userFeedback: userFeedback)
+    }
 
     @IBAction func crashOnException(_ sender: Any) {
         let userInfo: [String: String] = ["user-info-key-1": "user-info-value-1", "user-info-key-2": "user-info-value-2"]