+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - fix: Carthage for Xcode 12 #780
 - fix: Add missing SentrySdkInfo.h to umbrella header #779
 - ref: Remove event.json field #768
+- fix: Sanitize user info of captured errors #770
 
 ## 6.0.1